Rev 2-5-0 Ver A
I'm only doing version A until I've fully tested everything. And I
probably won't be able to test it for some time unfortunately.

This is a two layer board. I separated the traces for the audio and
video as best I could. This is fix issues with the noise channel causing
little bits of static to appear on screen. This was only noticeable in
Mega Man 2 with the black intro screen. But MM2 is one of my favorites,
so that is unnacceptable. If this doesn't work, I may have to do a 4
layer version.

Fixed all plated through holes. I think Altium likes to guess what you
are doing and set your default settings for you. I think what happened
was I started making the mounting holes which were unplated, and Altium
decided that every hole from then on should be unplated as well...

Moved components to bottom to make it easier to work on. Changed
regulator to surface mount, so that if I want to screw the multi-out
connector into the side of the chassis, the screw won't collide with the
old heatsink.

Added BOM with part numbers. I do not put part numbers for generic parts
like 0805 caps and resistors.